# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_eleven-am_voice-backend_internal_dto.APIKeyListResponse:
    properties:
      api_keys:
        items:
          $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.APIKeyResponse'
        type: array
    type: object
  github_com_eleven-am_voice-backend_internal_dto.APIKeyResponse:
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      expires_at:
        example: "2024-12-31T23:59:59Z"
        type: string
      id:
        example: key_abc123
        type: string
      last_used_at:
        example: "2024-01-20T15:45:00Z"
        type: string
      name:
        example: Production Key
        type: string
      prefix:
        example: sk-voice-abc
        type: string
    type: object
...
  github_com_eleven-am_voice-backend_internal_dto.AgentResponse:
    properties:
      active_installs:
        example: 500
        type: integer
      avg_rating:
        example: 4.5
        type: number
      capabilities:
        example:
        - search
        - weather
        items:
          type: string
        type: array
      category:
        example: productivity
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      description:
        example: A helpful voice assistant
        type: string
      developer_id:
        example: usr_xyz789
        type: string
      id:
        example: agt_abc123
        type: string
      is_public:
        example: true
        type: boolean
      is_verified:
        example: false
        type: boolean
      keywords:
        example:
        - helpful
        - assistant
        items:
          type: string
        type: array
      logo_url:
        example: https://example.com/logo.png
        type: string
      name:
        example: My Voice Agent
        type: string
      total_installs:
        example: 1000
        type: integer
      total_reviews:
        example: 100
        type: integer
      updated_at:
        example: "2024-01-20T15:45:00Z"
        type: string
    type: object
  github_com_eleven-am_voice-backend_internal_dto.CreateAPIKeyRequest:
    properties:
      expires_in_days:
        example: 90
        type: integer
      name:
        example: Production Key
        type: string
    type: object
  github_com_eleven-am_voice-backend_internal_dto.CreateAPIKeyResponse:
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      expires_at:
        example: "2024-12-31T23:59:59Z"
        type: string
      id:
        example: key_abc123
        type: string
      last_used_at:
        type: string
      name:
        example: Production Key
        type: string
      prefix:
        example: sk-voice-abc
        type: string
      secret:
        example: sk-voice-abcXXXXXXXXXXXXXXXXXXXXX
        type: string
    type: object
...
  github_com_eleven-am_voice-backend_internal_dto.UpdateAgentRequest:
    properties:
      capabilities:
        example:
        - new
        - capabilities
        items:
          type: string
        type: array
      category:
        enum:
        - productivity
        - education
        - entertainment
        - lifestyle
        - utility
        - social
        - finance
        - health
        - news
        - assistant
        example: education
        type: string
      description:
        example: Updated description
        type: string
      keywords:
        example:
        - updated
        - keywords
        items:
          type: string
        type: array
      logo_url:
        example: https://example.com/new-logo.png
        type: string
      name:
        example: Updated Agent Name
        type: string
    type: object
...
  github_com_eleven-am_voice-backend_internal_shared.APIError:
    properties:
      code:
        example: invalid_request
        type: string
      details:
        type: object
      message:
        example: Invalid request body
        type: string
    type: object
host: api.voice.example.com
info:
  contact: {}
  description: API server for voice agent platform
  title: Voice Backend API
  version: 1.0.0
paths:
...
  /agents/{id}:
    delete:
      description: Deletes an agent owned by the authenticated developer
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Delete an agent
      tags:
      - agents
    get:
      description: Returns a specific agent owned by the authenticated developer
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.AgentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Get an agent
      tags:
      - agents
    put:
      consumes:
      - application/json
      description: Updates an existing agent owned by the authenticated developer
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Agent update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.UpdateAgentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.AgentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Update an agent
      tags:
      - agents
...
  /apikeys:
    get:
      description: Returns all API keys owned by the authenticated developer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.APIKeyListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: List API keys
      tags:
      - apikeys
    post:
      consumes:
      - application/json
      description: Creates a new API key for the authenticated developer
      parameters:
      - description: API key details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.CreateAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Create an API key
      tags:
      - apikeys
  /apikeys/{id}:
    delete:
      description: Deletes an API key owned by the authenticated developer
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Delete an API key
      tags:
      - apikeys
...
securityDefinitions:
  APIKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  SessionAuth:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"